+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - Support for multiple sets of resource and operation defaults, including
   support for rules with 'resource' and 'op' expressions ([rhbz#1817547])
+- Support for "demote" value of resource operation's "on-fail" option
+  ([rhbz#1843079])
 
 ### Deprecated
 - `pcs resource [op] defaults <name>=<value>...` commands are deprecated now.
@@ -13,6 +15,7 @@
   several sets of defaults. ([rhbz#1817547])
 
 [rhbz#1817547]: https://bugzilla.redhat.com/show_bug.cgi?id=1817547
+[rhbz#1843079]: https://bugzilla.redhat.com/show_bug.cgi?id=1843079
 
 
 ## [0.10.6] - 2020-06-11